separated app from server
